Delivery_Project/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==4.4.0
djangorestframework==3.13.0
dnspython==2.6.1
geographiclib==2.0
geopy==2.3.0
grpcio==1.66.1
grpcio-tools==1.66.1
httpie==3.2.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
protobuf==5.28.0
Pygments==2.18.0
pymongo==4.7.3
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
